+++ common/api/__init__.py
@@ -3,7 +3,7 @@ import os
 import requests
 from datetime import datetime, timedelta
 from common.basedir import PERSIST
-from selfdrive.version import get_version
+from system.version import get_version
 
 API_HOST = os.getenv('API_HOST', 'https://api.commadotai.com')
