+++ selfdrive/athena/registration.py
@@ -12,7 +12,7 @@ from common.spinner import Spinner
 from common.basedir import PERSIST
 from selfdrive.controls.lib.alertmanager import set_offroad_alert
 from system.hardware import HARDWARE, PC
-from selfdrive.swaglog import cloudlog
+from system.swaglog import cloudlog
 
 
 UNREGISTERED_DONGLE_ID = "UnregisteredDevice"
