+++ selfdrive/boardd/pandad.py
@@ -11,7 +11,7 @@ from panda import DEFAULT_FW_FN, DEFAULT_H7_FW_FN, MCU_TYPE_H7, Panda, PandaDFU
 from common.basedir import BASEDIR
 from common.params import Params
 from system.hardware import HARDWARE
-from selfdrive.swaglog import cloudlog
+from system.swaglog import cloudlog
 
 
 def get_expected_signature(panda: Panda) -> bytes:
