+++ selfdrive/boardd/tests/boardd_old.py
@@ -13,7 +13,7 @@ import time
 
 import cereal.messaging as messaging
 from common.realtime import Ratekeeper
-from selfdrive.swaglog import cloudlog
+from system.swaglog import cloudlog
 from selfdrive.boardd.boardd import can_capnp_to_can_list
 from cereal import car
