+++ selfdrive/car/isotp_parallel_query.py
@@ -4,7 +4,7 @@ from functools import partial
 from typing import Optional
 
 import cereal.messaging as messaging
-from selfdrive.swaglog import cloudlog
+from system.swaglog import cloudlog
 from selfdrive.boardd.boardd import can_list_to_can_capnp
 from panda.python.uds import CanClient, IsoTpMessage, FUNCTIONAL_ADDRS, get_rx_addr_for_tx_addr
