+++ selfdrive/car/vin.py
@@ -6,7 +6,7 @@ import cereal.messaging as messaging
 import panda.python.uds as uds
 from panda.python.uds import FUNCTIONAL_ADDRS
 from selfdrive.car.isotp_parallel_query import IsoTpParallelQuery
-from selfdrive.swaglog import cloudlog
+from system.swaglog import cloudlog
 
 OBD_VIN_REQUEST = b'\x09\x02'
 OBD_VIN_RESPONSE = b'\x49\x02\x01'
