+++ selfdrive/controls/lib/lane_planner.py
@@ -3,7 +3,7 @@ from cereal import log
 from common.filter_simple import FirstOrderFilter
 from common.numpy_fast import interp
 from common.realtime import DT_MDL
-from selfdrive.swaglog import cloudlog
+from system.swaglog import cloudlog
 
 
 TRAJECTORY_SIZE = 33
