+++ selfdrive/loggerd/tests/test_loggerd.py
@@ -17,7 +17,7 @@ from common.params import Params
 from common.timeout import Timeout
 from selfdrive.loggerd.config import ROOT
 from selfdrive.manager.process_config import managed_processes
-from selfdrive.version import get_version
+from system.version import get_version
 from tools.lib.logreader import LogReader
 from cereal.visionipc import VisionIpcServer, VisionStreamType
 from common.transformations.camera import tici_f_frame_size, tici_d_frame_size, tici_e_frame_size
