+++ selfdrive/manager/build.py
@@ -11,8 +11,8 @@ from common.basedir import BASEDIR
 from common.spinner import Spinner
 from common.text_window import TextWindow
 from system.hardware import AGNOS
-from selfdrive.swaglog import cloudlog, add_file_handler
-from selfdrive.version import is_dirty
+from system.swaglog import cloudlog, add_file_handler
+from system.version import is_dirty
 
 MAX_CACHE_SIZE = 4e9 if "CI" in os.environ else 2e9
 CACHE_DIR = Path("/data/scons_cache" if AGNOS else "/tmp/scons_cache")
