+++ selfdrive/manager/process.py
@@ -16,7 +16,7 @@ from cereal import car
 from common.basedir import BASEDIR
 from common.params import Params
 from common.realtime import sec_since_boot
-from selfdrive.swaglog import cloudlog
+from system.swaglog import cloudlog
 from system.hardware import HARDWARE
 from cereal import log
