+++ selfdrive/sentry.py
@@ -6,8 +6,8 @@ from sentry_sdk.integrations.threading import ThreadingIntegration
 from common.params import Params
 from selfdrive.athena.registration import is_registered_device
 from system.hardware import HARDWARE, PC
-from selfdrive.swaglog import cloudlog
-from selfdrive.version import get_branch, get_commit, get_origin, get_version, \
+from system.swaglog import cloudlog
+from system.version import get_branch, get_commit, get_origin, get_version, \
                               is_comma_remote, is_dirty, is_tested_branch
