+++ selfdrive/test/helpers.py
@@ -4,7 +4,7 @@ from functools import wraps
 
 from system.hardware import PC
 from selfdrive.manager.process_config import managed_processes
-from selfdrive.version import training_version, terms_version
+from system.version import training_version, terms_version
 
 
 def set_params_enabled():
