+++ selfdrive/test/test_onroad.py
@@ -39,7 +39,7 @@ PROCS = {
   "./_soundd": 1.0,
   "selfdrive.monitoring.dmonitoringd": 1.90,
   "./proclogd": 1.54,
-  "selfdrive.logmessaged": 0.2,
+  "system.logmessaged": 0.2,
   "./clocksd": 0.02,
   "./ubloxd": 0.02,
   "selfdrive.tombstoned": 0,
