+++ selfdrive/thermald/fan_controller.py
@@ -3,7 +3,7 @@ from abc import ABC, abstractmethod
 
 from common.realtime import DT_TRML
 from common.numpy_fast import interp
-from selfdrive.swaglog import cloudlog
+from system.swaglog import cloudlog
 from selfdrive.controls.lib.pid import PIDController
 
 class BaseFanController(ABC):
