+++ selfdrive/thermald/power_monitoring.py
@@ -5,7 +5,7 @@ from cereal import log
 from common.params import Params, put_nonblocking
 from common.realtime import sec_since_boot
 from system.hardware import HARDWARE
-from selfdrive.swaglog import cloudlog
+from system.swaglog import cloudlog
 from selfdrive.statsd import statlog
 
 CAR_VOLTAGE_LOW_PASS_K = 0.091 # LPF gain for 5s tau (dt/tau / (dt/tau + 1))
